backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.1
pydantic==2.5.0
pydantic[email]==2.5.0
redis==5.0.1
